Add OrthographicCamera

+++ src/geometry/Sphere.cpp
@@ -8,14 +8,14 @@ bool Sphere::intersect(const Ray &ray, std::vector<Vector3> &intersections) {
     Vector3 rayDirection(ray.direction.normalize());
 
     Vector3 oc = ray.origin - center;
-    float a = rayDirection.dot(rayDirection);
-    float b = 2 * oc.dot(rayDirection);
-    float c = oc.dot(oc) - radius * radius;
+    double a = rayDirection.dot(rayDirection);
+    double b = 2 * oc.dot(rayDirection);
+    double c = oc.dot(oc) - radius * radius;
     double delta = b * b - 4 * a * c;
 
     if (delta > 0) {
-        float f0 = (-b - sqrt(delta)) / (2.0 * a);
-        float f1 = (-b + sqrt(delta)) / (2.0 * a);
+        double f0 = (-b - sqrt(delta)) / (2.0 * a);
+        double f1 = (-b + sqrt(delta)) / (2.0 * a);
 
         if(f0 >= 0) {
             intersections.push_back(f0 * rayDirection + ray.origin);
@@ -27,7 +27,7 @@ bool Sphere::intersect(const Ray &ray, std::vector<Vector3> &intersections) {
         return !intersections.empty();
     }
     if (delta == 0) {
-        float f = (-b / (2.0 * a));
+        double f = (-b / (2.0 * a));
 
         if(f > 0) {
             intersections.push_back(f * rayDirection + ray.origin);
